+++ app/api/auth_routes.py
@@ -2,7 +2,7 @@ from flask import Blueprint, session, request
 from app.models import User, db, ClanUsers
 from app.forms import LoginForm, SignUpForm
 from flask_login import current_user, login_user, logout_user
-from user_repository import create_user
+from app.api.user.user_repository import user_repository
 
 auth_routes = Blueprint('auth', __name__)
 
@@ -34,7 +34,7 @@ def authenticate():
     """
     if current_user.is_authenticated:
         return current_user.to_dict()
-    return {'errors': ['Unauthorized']}
+    return {'errors': ['Unauthorized']}, 403
 
 
 @auth_routes.route('/login', methods=['POST'])
@@ -47,10 +47,7 @@ def login():
     # form manually to validate_on_submit can be used
     form['csrf_token'].data = request.cookies['csrf_token']
     if form.validate_on_submit():
-        user = User.query.filter(User.email == form.data['email']).first()
-        is_clan_member = clan_member_check(user.id)
-        if is_clan_member:
-            user = User.query.filter(User.email == form.data['email']).join(ClanUsers).first()
+        user = user_repository.login(form.data['email'], form.data['password'])
         # Add the user to the session, we are logged in!
         login_user(user)
         return user.to_dict()
@@ -74,13 +71,13 @@ def sign_up():
     form = SignUpForm()
     form['csrf_token'].data = request.cookies['csrf_token']
     if form.validate_on_submit():
-        user = create_user(
+        user = user_repository.create_user(
             username=form.data['username'],
             email=form.data['email'],
             password=form.data['password']
         )
         login_user(user)
-        return user
+        return user.to_dict()
     return {'errors': validation_errors_to_error_messages(form.errors)}, 401
