+++ app/api/user/__init__.py
@@ -1,0 +1,1 @@
+from .user_repository import user_repository
